fix: 좋아요 등록 및 취소 메서드 분기 처리

+++ src/main/java/com/gaethering/gaetheringserver/domain/board/service/HeartServiceImpl.java
@@ -3,8 +3,6 @@ package com.gaethering.gaetheringserver.domain.board.service;
 import com.gaethering.gaetheringserver.domain.board.dto.HeartResponse;
 import com.gaethering.gaetheringserver.domain.board.entity.Heart;
 import com.gaethering.gaetheringserver.domain.board.entity.Post;
-import com.gaethering.gaetheringserver.domain.board.exception.AlreadyPushHeartException;
-import com.gaethering.gaetheringserver.domain.board.exception.HeartNotFoundException;
 import com.gaethering.gaetheringserver.domain.board.exception.PostNotFoundException;
 import com.gaethering.gaetheringserver.domain.board.repository.HeartRepository;
 import com.gaethering.gaetheringserver.domain.board.repository.PostRepository;
@@ -14,6 +12,8 @@ import com.gaethering.gaetheringserver.domain.member.repository.member.MemberRep
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -33,41 +33,29 @@ public class HeartServiceImpl implements HeartService {
         Member member = memberRepository.findByEmail(email)
                 .orElseThrow(() -> new MemberNotFoundException());
 
-        if (heartRepository.existsByPostAndMember(post, member)) {
-            throw new AlreadyPushHeartException();
+        Optional<Heart> optionalHeart = heartRepository.findByPostAndMember(post, member);
+
+        if (optionalHeart.isPresent()) {
+            Heart heart = optionalHeart.get();
+
+            heartRepository.delete(heart);
+            post.cancelPostHeart(heart);
+
+            return HeartResponse.builder()
+                    .memberId(member.getId())
+                    .postId(post.getId())
+                    .likeCnt(heartRepository.countByPost(post).intValue())
+                    .build();
         }
 
-        Heart heart = Heart.builder()
+        Heart newHeart = Heart.builder()
                 .post(post)
                 .member(member)
                 .build();
 
-        heartRepository.save(heart);
+        heartRepository.save(newHeart);
 
-        heart.mappingPost(post);
-
-        return HeartResponse.builder()
-                .memberId(member.getId())
-                .postId(post.getId())
-                .likeCnt(heartRepository.countByPost(post).intValue())
-                .build();
-    }
-
-    @Override
-    @Transactional
-    public HeartResponse cancelHeart(Long postId, String email) {
-
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new PostNotFoundException());
-
-        Member member = memberRepository.findByEmail(email)
-                .orElseThrow(() -> new MemberNotFoundException());
-
-        Heart heart = heartRepository.findByPostAndMember(post, member)
-                .orElseThrow(() -> new HeartNotFoundException());
-
-        heartRepository.delete(heart);
-        post.cancelPostHeart(heart);
+        newHeart.mappingPost(post);
 
         return HeartResponse.builder()
                 .memberId(member.getId())
